Provide a mechanism for generating entire tables (#7)
* Provide a mechanism for generating entire tables

This adds functions `collect_tables`, which turns the output of
`generate` into a `NamedTuple` of tables, and `generate_tables` which is
just a convenience atop `collect_tables(generate(...))`.

* Update the documentation

The README was updated for the change to include an `rng` argument in
method definitions but the package docs were not. This commit makes the
example runnable. It also replaces the `OrderedDict` example with
`collect_tables`.

* Bump package version to 0.2.1